requirements.txt
openai==1.93.0
python-dotenv==1.0.0
discord.py==2.3.1